pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ypt"
version = "0.1.0"
description = "Yoda Performance Testing (ypt) suite"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "GPL-3.0-or-later"}

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "python-irodsclient==3.1.1",
    "locust==2.43.3",
    "urllib3==2.6.3",
    "requests==2.32.4",
    "webdavclient3==3.14.7",
    "matplotlib==3.10.8",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "ruff==0.15.5",
    "flake8==7.3.0",
    "flake8-import-order==0.19.1",
    "flake8-pyproject==1.2.4",
    "darglint==1.8.1",
    "mypy==1.19.1",
]

[tool.setuptools]
packages = ["ypt"]

[project.scripts]
yoda-performance-tests = "ypt.ypt:main"

[tool.ruff]
exclude = [
    ".idea",
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
    ".mypy_cache",
    ".ruff_cache",
    "ypt"
]
line-length = 127
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "C90", "FURB", "N", "UP", "YTT", "PLE", "FURB", "SIM", "B", "C4", "ISC", "T20"]
ignore = ["T201"]

[tool.flake8]
ignore = ["E402", "DAR"]
import-order-style = "smarkets"
strictness = "short"
docstring_style = "sphinx"
max-line-length = 127
exclude = ["venv", "build"]
application-import-names=["lf_irods"]

[tool.mypy]
strict = true
exclude = ["venv", "build", "ypt"]

ignore_missing_imports = true
warn_unreachable = true
no_implicit_optional = true
check_untyped_defs = false
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_defs = true
disable_error_code = ["var-annotated", "type-arg", "union-attr", "untyped-decorator", "misc"]
show_error_codes = true
show_error_context = true
pretty = true
